Slipstream the appropriate backend host for the environment
When downloading a zip from the page, the `host.lua` file was being
included statically from the file on disk. This means the "dev"
version of the mod connected to the production backend.

This change adds host.lua to the slipstream list, supplying whatever
the contents of the BACKEND env variable is, or defaulting to the
production host.