Convert the backend code to Typescript
Introduced an abstraction between the application code and
the database that relies on Typebox for schema validation
and runtypes. The goal is to be able to switch away from
MongoDB.

The Mongoose schema had typed user ids as "Number", but this
is unsafe with Twitch; Twitch's REST endpoints give ids
as strings because they are planning for the potential for
ids to exceed safe JSON numerical values. We should be
typing and treating them as either bigints or strings
for this reason. Mongoose didn't complain because it
implicitly coerces the values. Mongoose doesn't seem to
like bigint values, so for now I'm converting the ids
to strings when interacting with the DB and back to
bigints in the application itself.